+++ test/unit/fuzz-tools.test.js
@@ -317,6 +317,7 @@ import * as outerbridgeV337 from '../../lib/outerbridge-v337.js';
 import * as icrsV338 from '../../lib/icrs-v338.js';
 import * as cormackLehaneV339 from '../../lib/cormack-lehane-v339.js';
 import * as clarkLevelV340 from '../../lib/clark-level-v340.js';
+import * as masonRadialHeadV341 from '../../lib/mason-radial-head-v341.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -579,6 +580,7 @@ const MODULES = {
   'icrs-v338.js': icrsV338,
   'cormack-lehane-v339.js': cormackLehaneV339,
   'clark-level-v340.js': clarkLevelV340,
+  'mason-radial-head-v341.js': masonRadialHeadV341,
   'tb-testing.js': tbTesting,
 };